+++ frontend/src/routes/Router.tsx
@@ -26,7 +26,7 @@ import EditEvent from '../pages/EditEvent';
 import AddAnnouncement from '../pages/AddAnnouncement';
 import EditAnnouncement from '../pages/EditAnnouncement';
 import Announcement from '../pages/Announcement';
-import ActivityLog from '../pages/ActivityLog';
+import ActivityLogs from '../pages/ActivityLogs';
 import AddActivityLog from '../pages/AddActivityLog';
 
 const Router = () => {
@@ -70,7 +70,7 @@ const Router = () => {
             <Route path="add" element={<AddEvent />} />
           </Route>
           <Route path="activity-log" element={<Outlet />}>
-            <Route index element={<ActivityLog />} />
+            <Route index element={<ActivityLogs />} />
             <Route path="add" element={<AddActivityLog />} />
           </Route>
         </Route>
